requirements.txt
streamlit
langchain
openai
faiss-cpu
sentence-transformers
python-docx
pymupdf
pandas
python-dotenv
scikit-learn

